requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
analytics-python==1.4.post1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==1.10.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
configparser==5.0.2
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
exceptiongroup==1.1.3
fastapi==0.101.1
ffmpy==0.3.1
filelock==3.12.2
Flask==2.3.2
Flask-CacheBuster==1.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.2
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
gradio==3.40.1
gradio-client==0.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
linkify-it-py==2.0.2
markdown-it-py==3.0.0
markdown2==2.4.10
MarkupSafe==2.1.3
matplotlib==3.7.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.4
opencv-python==4.8.0.76
orjson==3.9.5
packaging==23.1
pandas==2.0.3
paramiko==3.3.1
Pillow==8.2.0
pkgutil-resolve-name==1.3.10
protobuf==4.24.0
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==12.0.1
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.2.0
pydantic-core==2.6.0
pydeck==0.8.0
pydub==0.25.1
Pygments==2.16.1
Pympler==1.0.1
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.10.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
streamlit==1.25.0
sympy==1.12
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uc-micro-py==1.0.2
ultralytics==8.0.157
urllib3==2.0.4
uvicorn==0.23.2
validators==0.21.2
watchdog==3.0.0
websockets==11.0.3
werkzeug==2.3.7
yarl==1.9.2
zipp==3.16.2
